Added a backoff mechanism with failureThreshold to confirm ownerChecks failure.
Added a CLI flag .

+++ pkg/server/backuprestoreserver.go
@@ -66,7 +66,7 @@ func NewBackupRestoreServer(logger *logrus.Logger, config *BackupRestoreComponen
 	var ownerChecker common.Checker
 	if occ.OwnerName != "" && occ.OwnerID != "" {
 		resolver := common.NewCachingResolver(net.DefaultResolver, clock.RealClock{}, occ.OwnerCheckDNSCacheTTL.Duration)
-		ownerChecker = common.NewOwnerChecker(occ.OwnerName, occ.OwnerID, occ.OwnerCheckTimeout.Duration, resolver, serverLogger)
+		ownerChecker = common.NewOwnerChecker(occ.OwnerName, occ.OwnerID, occ.OwnerCheckTimeout.Duration, resolver, serverLogger, occ.OwnerCheckFailureThreshold)
 	}
 	etcdProcessKiller := common.NewNamedProcessKiller(config.EtcdProcessName, common.NewGopsutilProcessLister(), serverLogger)
 	defragmentationSchedule, err := cron.ParseStandard(config.DefragmentationSchedule)
@@ -275,10 +275,6 @@ func (b *BackupRestoreServer) runEtcdProbeLoopWithSnapshotter(ctx context.Contex
 			}
 		}
 
-		// set server's healthz endpoint status to OK so that
-		// etcd is marked as ready to serve traffic
-		handler.SetStatus(http.StatusOK)
-
 		if runServerWithSnapshotter {
 			if b.ownerChecker != nil {
 				// Check if the actual owner ID matches the expected one
@@ -317,6 +313,10 @@ func (b *BackupRestoreServer) runEtcdProbeLoopWithSnapshotter(ctx context.Contex
 					continue
 				}
 			}
+
+			// set server's healthz endpoint status to OK so that
+			// etcd is marked as ready to serve traffic
+			handler.SetStatus(http.StatusOK)
 
 			// The decision to either take an initial delta snapshot or
 			// or a full snapshot directly is based on whether there has
@@ -434,18 +434,27 @@ func (b *BackupRestoreServer) runEtcdProbeLoopWithSnapshotter(ctx context.Contex
 				// Stop owner check watchdog
 				ownerCheckWatchdog.Stop()
 
-				// If the owner check fails or returns false, kill the etcd process
-				// to ensure that any open connections from kube-apiserver are terminated
-				result, err := b.ownerChecker.Check(ctx)
-				if err != nil || !result {
-					if _, err := b.etcdProcessKiller.Kill(ctx); err != nil {
-						b.logger.Errorf("Could not kill etcd process: %v", err)
+				exponentialBackoff := backoff.NewExponentialBackOffConfig(b.config.ExponentialBackoffConfig.AttemptLimit, b.config.ExponentialBackoffConfig.Multiplier, b.config.ExponentialBackoffConfig.ThresholdTime.Duration)
+				isFail := ownerCheckWatchdog.Confirm(ctx, b.config.OwnerCheckConfig.OwnerCheckFailureThreshold, exponentialBackoff)
+
+				if isFail {
+					// If the owner check fails or returns false, kill the etcd process
+					// to ensure that any open connections from kube-apiserver are terminated
+					result, err := b.ownerChecker.Check(ctx)
+					if err != nil || !result {
+						if _, err := b.etcdProcessKiller.Kill(ctx); err != nil {
+							b.logger.Errorf("Could not kill etcd process: %v", err)
+						}
 					}
 				}
 			}
 
 		} else {
 			// for the case when snapshotter is not configured
+
+			// set server's healthz endpoint status to OK so that
+			// etcd is marked as ready to serve traffic
+			handler.SetStatus(http.StatusOK)
 			<-ctx.Done()
 			handler.SetStatus(http.StatusServiceUnavailable)
 			b.logger.Infof("Received stop signal. Terminating !!")
